Ensure null hostility
If an empty element was presented to a scalar element handler, the
element handler's onCharacters method was never called. This allowed
scalar element handlers to return `null`. Now, if the handler has
not been called for an element, it will be called explicitly with an
empty string before returning.

Fix: https://github.com/io7m/blackthorne/issues/4